+++ gemma-core/src/main/java/ubic/gemma/persistence/service/analysis/expression/diff/DifferentialExpressionAnalysisDao.java
@@ -20,16 +20,14 @@ package ubic.gemma.persistence.service.analysis.expression.diff;
 
 import ubic.gemma.model.analysis.SingleExperimentAnalysis;
 import ubic.gemma.model.analysis.expression.diff.DifferentialExpressionAnalysis;
-import ubic.gemma.model.analysis.expression.diff.DifferentialExpressionAnalysisValueObject;
 import ubic.gemma.model.expression.experiment.BioAssaySet;
 import ubic.gemma.model.expression.experiment.ExperimentalFactor;
 import ubic.gemma.model.expression.experiment.ExpressionExperiment;
+import ubic.gemma.model.expression.experiment.FactorValue;
 import ubic.gemma.model.genome.Gene;
-import ubic.gemma.model.genome.Taxon;
 
 import javax.annotation.Nullable;
 import java.util.Collection;
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -38,23 +36,40 @@ import java.util.Map;
 public interface DifferentialExpressionAnalysisDao extends ubic.gemma.persistence.service.analysis.AnalysisDao<DifferentialExpressionAnalysis> {
 
     /**
-     * Find by associated experiment via {@link SingleExperimentAnalysis#getExperimentAnalyzed()}.
-     * @param includeSubSets include subsets of the specified experiments
+     * Find by associated {@link ubic.gemma.model.expression.experiment.ExpressionExperiment}.
+     *
+     * @param includeSubSets include subsets of the specified experiment
      */
-    Collection<DifferentialExpressionAnalysis> findByExperimentAnalyzed( ExpressionExperiment experiment, boolean includeSubSets );
+    Collection<DifferentialExpressionAnalysis> findByExperiment( ExpressionExperiment experiment, boolean includeSubSets );
 
+    /**
+     * Find by associated experiment via {@link SingleExperimentAnalysis#getExperimentAnalyzed()}.
+     */
     Collection<DifferentialExpressionAnalysis> findByExperimentAnalyzed( BioAssaySet experimentAnalyzed );
 
     /**
-     * Given a collection of experiments returns a Map of Analysis --&gt; collection of Experiments
-     * The collection of investigations returned by the map will include all the investigations for the analysis key iff
-     * one of the investigations for that analysis was in the given collection started with
+     * Find by associated {@link ExpressionExperiment}s.
      *
-     * @param experiments    experiments
      * @param includeSubSets include subsets of the specified experiments
-     * @return map to analyses
+     * @return map to analyses grouped by source experiment
      */
-    Map<BioAssaySet, Collection<DifferentialExpressionAnalysis>> findByExperimentsAnalyzed( Collection<ExpressionExperiment> experiments, boolean includeSubSets );
+    Map<ExpressionExperiment, Collection<DifferentialExpressionAnalysis>> findByExperiments( Collection<ExpressionExperiment> experiments, boolean includeSubSets );
+
+    /**
+     * Retrieve analysis for the given experiment or experiment subset IDs.
+     * <p>
+     * Results are grouped by source experiment.
+     *
+     * @param experimentIds   experiment IDs (no subset allowed!)
+     * @param includeSubSets  include analyses for subsets of the requested experiment IDs
+     * @param arrayDesignUsed a map of experiments or subset ID to array design IDs used
+     * @param ee2fv           a map of experiments or subset ID to factor values used
+     * @return map to analyses grouped by source experiment
+     */
+    Map<ExpressionExperiment, Collection<DifferentialExpressionAnalysis>> findByExperimentIds(
+            Collection<Long> experimentIds, boolean includeSubSets,
+            @Nullable Map<Long, Collection<Long>> arrayDesignUsed,
+            @Nullable Map<Long, Collection<FactorValue>> ee2fv );
 
     Collection<DifferentialExpressionAnalysis> findByName( String name );
 
@@ -72,22 +87,12 @@ public interface DifferentialExpressionAnalysisDao extends ubic.gemma.persistenc
      * Find an analysis for the given experiment and identifier.
      */
     @Nullable
-    DifferentialExpressionAnalysis findByExperimentAndAnalysisId( ExpressionExperiment experimentAnalyzed, Long analysisId, boolean includeSubSets );
+    DifferentialExpressionAnalysis findByExperimentAndAnalysisId( ExpressionExperiment experimentAnalyzed, boolean includeSubSets, Long analysisId );
 
-    Map<Long, Collection<DifferentialExpressionAnalysis>> findByExperimentAnalyzedId( Collection<Long> experimentAnalyzedId );
-
+    /**
+     * Find experiments (or subsets) that have differential expression analyses for the given gene.
+     */
     Collection<BioAssaySet> findExperimentsWithAnalyses( Gene gene );
 
     Collection<Long> getExperimentsWithAnalysis( Collection<Long> eeIds, boolean includeSubSets );
-
-    Collection<Long> getExperimentsWithAnalysis( Taxon taxon );
-
-    /**
-     * Retrieve analysis VOs for the given experiment IDs.
-     *
-     * @param includeSubSets include analyses for subsets of the requested experiment IDs
-     * @param includeAssays  include analyzed assays in the value object (can be large)
-     */
-    Map<Long, List<DifferentialExpressionAnalysisValueObject>> getAnalysesByExperimentIds(
-            Collection<Long> expressionExperimentIds, int offset, int limit, boolean includeSubSets, boolean includeAssays );
 }