+++ gemma-rest/src/main/resources/restapidocs/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Updates
 
+### Update 2.9.3
+
+Add various single-cell metrics to the `BioAssayValueObject` model.
+
+Ignore the `offset` and `limit` parameters of `getDatasetDifferentialExpressionAnalyses` and schedule it for removal in
+the 2.10 release. This parameter was not really necessary given the few number of analyses per dataset and was causing
+issues with the pagination over the result sets.
+
 ### Update 2.9.2
 
 Subject, predicates and objects of statements are now exposed in the `FilterArgExpressionExperiment` model. Their URIs